+++ backend/lost/api/sia/endpoint.py
@@ -168,3 +168,56 @@ class Configuration(Resource):
             print ('Anno task config in endpoint', re)
             dbm.close_session()
             return re
+
+@namespace.route('/review')
+class Review(Resource):
+    @jwt_required 
+    def post(self):
+        dbm = access.DBMan(LOST_CONFIG)
+        identity = get_jwt_identity()
+        user = dbm.get_user_by_id(identity)
+        if not user.has_role(roles.DESIGNER):
+            dbm.close_session()
+            return "You need to be {} in order to perform this request.".format(roles.DESIGNER), 401
+
+        else:
+            data = json.loads(request.data)
+            re = sia.review(dbm, data, user.idx, DATA_URL)
+            dbm.close_session()
+            return re
+
+
+@namespace.route('/reviewoptions/<int:pe_id>')
+@namespace.param('pe_id', 'The id of reviewed pipe element.')
+class ReviewOptions(Resource):
+    @jwt_required 
+    def get(self, pe_id):
+        dbm = access.DBMan(LOST_CONFIG)
+        identity = get_jwt_identity()
+        user = dbm.get_user_by_id(identity)
+        if not user.has_role(roles.DESIGNER):
+            dbm.close_session()
+            return "You need to be {} in order to perform this request.".format(roles.DESIGNER), 401
+
+        else:
+            re = sia.reviewoptions(dbm, pe_id, user.idx)
+            dbm.close_session()
+            return re
+
+@namespace.route('/reviewupdate/<int:pe_id>')
+@namespace.param('pe_id', 'The id of reviewed pipe element.')
+class ReviewUpdate(Resource):
+    @jwt_required 
+    def post(self, pe_id):
+        dbm = access.DBMan(LOST_CONFIG)
+        identity = get_jwt_identity()
+        user = dbm.get_user_by_id(identity)
+        if not user.has_role(roles.DESIGNER):
+            dbm.close_session()
+            return "You need to be {} in order to perform this request.".format(roles.DESIGNER), 401
+
+        else:
+            data = json.loads(request.data)
+            re = sia.review_update(dbm, data, user.idx, pe_id)
+            dbm.close_session()
+            return re