+++ frontend/lost/src/types/index.js
@@ -7,5 +7,6 @@ import * as mia from './mia'
 import * as sia from './sia'
 import * as worker from './worker'
 import * as pipeline from './pipeline'
+import * as siaReview from './siaReview'
 
-export default { ...auth, ...label, ...group, ...user, ...annoTask, ...mia, ...sia, ...worker, ...pipeline}
+export default { ...auth, ...label, ...group, ...user, ...annoTask, ...mia, ...sia, ...worker, ...pipeline, ...siaReview}