+++ frontend/lost/src/views/SiaReview.js
@@ -7,27 +7,27 @@ import {
     Row
 } from 'reactstrap'
 
-
+import SIAReview from '../components/SIAReview/SIAReview.js'
 
 class SiaReview extends Component {
 	render(){
-		return (
-			<Row >
-				<Col>
-					<Card>
-						<CardHeader>
-							Single Image Annotation - Review !
-						</CardHeader>
-						<CardBody >
-							<Row>
-								<Col xs='12'>
-								Review here ...
-								</Col>
-							</Row>
-						</CardBody> 
-					</Card>
-				</Col>
-			</Row>
+		return (<SIAReview/>
+			// <Row >
+			// 	<Col>
+			// 		<Card>
+			// 			<CardHeader>
+			// 				Single Image Annotation - Review !
+			// 			</CardHeader>
+			// 			<CardBody >
+			// 				{/* <Row> */}
+			// 					{/* <Col xs='12'> */}
+			// 					<SIAReview></SIAReview>
+			// 					{/* </Col> */}
+			// 				{/* </Row> */}
+			// 			</CardBody> 
+			// 		</Card>
+			// 	</Col>
+			// </Row>
 		)
 	}
 }