+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - Unreleased
+### Added
+- SIAReview:
+  * Review tool for annotation tasks in designer interfaces
+  * Allows to view and adjust annotations for SIA and MIA tasks
+  * Allows also to filter annotations by iteration 
+
 ## [1.1.4] -  2020-06-08
 ### Fixed
 - Updated nodejs version in oder to get ci running