+++ frontend/lost/src/components/SIA/lost-sia/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2020-06-20
+### Fixed
+- Fixed multiple image load events for the same image, that lead to wrong svg sizes
+
 ## [0.2.0] - 2020-06-05
 ### Fixed 
 - Added camera move on wasd keys